requirements.txt
langchain==0.2.1
langchain-core==0.2.2
langchain-community==0.2.1
langchain-groq==0.1.5
langgraph==0.0.69
chromadb==0.4.24
sentence-transformers==2.7.0
faiss-cpu==1.8.0
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.6.4
python-dotenv==1.0.1
streamlit==1.39.0
pandas==2.2.3
numpy==1.26.4
groq==0.9.0
PyPDF2